pyproject.toml
[project]
name = "cpv-cot"
version = "0.1.0"
description = "CPV-CoT: Cross-Perspective Verification Chain-of-Thought"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "datasets>=2.14.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_backend"